bus: move type
